Release buzz 2.0.0
This change includes few but notable changes:

- A significant upgrade to the `imap` crate.
- Significantly updated parsing for buzz config files (#28),
  which includes support for user-configured icons.
- Support for multiple folders (#24).
- A complete swap of the library used for tray icons (#25).

From a user's perspective, the main change is that config files need to
be updated to match the new format. In particular, this means changing
`accounts` from a set of named tables to an array of tables, and moving
from `folder = "something"` to `folders = ["something"]`. See the README
for an updated example config.

Big thanks to @DanielVoogsgerd for the changes in this release!